Better organized devices and definitions
