+++ Assets/MixedRealityToolkit/_Core/Definitions/SpatialAwarenessSystem.meta
