+++ Assets/MixedRealityToolkit/_Core/Definitions/SpatialAwarenessSystem/SpatialAwarenessMeshLevelOfDetail.cs
