+++ Assets/MixedRealityToolkit/_Core/Definitions/SpatialAwarenessSystem/SpatialMeshDisplayOptions.cs
