+++ Assets/MixedRealityToolkit/_Core/Definitions/SpatialAwarenessSystem/SpatialMeshObject.cs
