+++ Assets/MixedRealityToolkit/_Core/Interfaces/Devices/SpatialObservers.meta
@@ -1,0 +1,8 @@
+fileFormatVersion: 2
+guid: 2049c9659a6849847bb17898b460b437
+folderAsset: yes
+DefaultImporter:
+  externalObjects: {}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
